% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commits.R
\name{get_commits}
\alias{get_commits}
\title{get df of recruits based on certain paramters}
\usage{
get_commits(college, years)
}
\arguments{
\item{college}{college name (to avoid confusion like Southern Cal and South Carolina, which would both be USC)}

\item{year}{the year of the recruits (it works with any year 247 has data for)}
}
\value{
data.frame
}
\description{
get df of recruits based on certain paramters
}
