% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hs_players.R
\name{get_hs_players}
\alias{get_hs_players}
\title{get df of hs players based on certain paramters}
\usage{
get_hs_players(state, years)
}
\arguments{
\item{state}{state abbreviation}

\item{year}{HS grad year}
}
\value{
data.frame of players from that state
}
\description{
get df of hs players based on certain paramters
}
