% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_position.R
\name{get_position}
\alias{get_position}
\title{get df of players based on positions}
\usage{
get_position(position, year)
}
\arguments{
\item{position}{position of players}

\item{year}{HS grad year}
}
\value{
data.frame of players of that position
}
\description{
get df of players based on positions
}
