% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_of_players.R
\name{map_of_players}
\alias{map_of_players}
\title{displays a leaflet map of all selected players in DF}
\usage{
map_of_players(players)
}
\arguments{
\item{players}{a data.frame of players, has to have Hometown column.}
}
\value{
nothing, prints a leaflet
}
\description{
displays a leaflet map of all selected players in DF
}
