% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hs_players.R, R/get_position.R
\name{simpleCap}
\alias{simpleCap}
\title{helper function to format committs}
\usage{
simpleCap(x)

simpleCap(x)
}
\arguments{
\item{school}{name of school}
}
\value{
string of corrected school name

string of corrected school name
}
\description{
helper function to format committs

helper function to format committs
}
